# pgtype/hstore: Make text parsing about 6X faster