@evanj 2023-06-15
Feel free to reject this: This is a big change. I would be happy to attempt to make this change in multiple smaller pieces if that makes sense. I also tested this with a really gross fuzz test that round-trips values to postgres. I would be happy to try to contribute that if it would be helpful, but I will need to clean it up.