@jackc 2023-06-16
> using string.IndexByte to scan double quoted strings: it has architecture-specific assembly implementations, and most of the time is spent in key/value strings.

and

> parsing directly into the Hstore, rather than copying into it.

seem like pretty obvious wins. Also, sizing the map appropriately upfront.

I'm not so sure about the single `string` and `[]string` to reduce allocations. I've used that technique multiple times in pgx and it yielded impressive results. But I've also reverted it multiple times because on rare occasions people ended up pinning large amounts of memory unexpectedly. And that type issue can be difficult to diagnose.

If it would make sense to separate that out, then the rest should be mergeable. But it may or may not be worth the work without the big win from the reduced allocations.

I guess the same issue would apply to #1646.

---

Also, if you want to make a "fasterhstore" package with the memory optimizations I'd be happy to include a link to it in the README.md.