@evanj 2023-06-16
Totally reasonable! Let me remove those parts of the changes.

Good idea about the hstore alternative. I was already considering it, but I think it would be even better if it was `map[string]pgtype.Text` since that should remove the need for the `[]string` since it gets rid of the pointers.