@evanj 2023-06-16
Now that I looked at the code again, I am going to undo the single string part, but I don't think this is actually a change for text parsing. The text parsing code *already* allocates a single backing slice. I learned this because my first attempt to parse directly into the map was slower, and I was very confused about why.

In the current master branch:

1. `parseHstore` allocates a slice of `[]Text` by building it with `append()`: https://github.com/jackc/pgx/blob/master/pgtype/hstore.go#L395
2. `scanString` converts each one of these elements to *string by taking a pointer to it, and storing that in the Hstore map: https://github.com/jackc/pgx/blob/master/pgtype/hstore.go#L258

I'm pretty sure this means the backing `[]Text` won't be deallocated until all the `*string` pointers in the Hstore go away, which is the *same* as this current text parser.

Since this is *not* a change, I would argue this part of the text parser is okay.

Note that this *is* a change for binary parsing, but if this part is okay for text parsing, I would argue it should also be okay for binary parsing, (or I should remove it from both).

... change coming shortly.